Updating heading/sub heading to match latest HMRC pattern
Latest pattern can be found here - https://design.tax.service.gov.uk/hmrc-design-patterns/page-heading/

This update leads to improved accessibility, preventing the sub heading and the hidden content within the sub heading being read out as part of H1 and and when H1 is within the legend of a fieldset. Reducing repetition an cognitive overload for assistive tech users.